Extract SearchEngine and SearchUi modules
