+++ assets/js/search.js
@@ -2,52 +2,8 @@
 
 'use strict';
 
-var lunr = require('lunr');
-var querystring = require('querystring');
-var url = require('url');
-
-var SEARCH_INPUT_ID = 'search-input';
-var SEARCH_RESULTS_ID = 'search-results';
-
-function fetchIndex(baseUrl) {
-  return new Promise(function(resolve, reject) {
-    var req = new XMLHttpRequest(),
-        indexUrl = baseUrl + '/search-index.json';
-
-    req.addEventListener('load', function() {
-      var rawJson;
-
-      try {
-        rawJson = JSON.parse(this.responseText);
-        resolve({
-          urlToDoc: rawJson.urlToDoc,
-          index: lunr.Index.load(rawJson.index)
-        });
-      } catch (err) {
-        reject(new Error('failed to parse ' + indexUrl));
-      }
-    });
-    req.open('GET', indexUrl);
-    req.send();
-  });
-}
-
-function parseSearchQuery(queryUrl) {
-  return querystring.parse(url.parse(queryUrl).query).q;
-}
-
-function getResults(query, searchIndex) {
-  var results = searchIndex.index.search(query);
-
-  results.forEach(function(result) {
-    var urlAndTitle = searchIndex.urlToDoc[result.ref];
-
-    Object.keys(urlAndTitle).forEach(function(key) {
-      result[key] = urlAndTitle[key];
-    });
-  });
-  return results;
-}
+var SearchEngine = require('./search-engine');
+var SearchUi = require('./search-ui');
 
 function writeResults(searchQuery, doc, searchBox, resultsList, results) {
   if (searchQuery) {
@@ -71,39 +27,12 @@ function writeResults(searchQuery, doc, searchBox, resultsList, results) {
   });
 }
 
-// eslint-disable-next-line
-// based on https://github.com/angular/angular.js/blob/54ddca537/docs/app/src/search.js#L198-L206
-function SearchUi(doc, inputElement) {
-  var isForwardSlash = function(keyCode) {
-    return keyCode === 191;
-  };
-
-  var isInput = function(el) {
-    return el.tagName.toLowerCase() === 'input';
-  };
-
-  var giveSearchFocus = function() {
-    inputElement.focus();
-  };
-
-  var onKeyDown = function(event) {
-    if (isForwardSlash(event.keyCode) && !isInput(doc.activeElement)) {
-      event.stopPropagation();
-      event.preventDefault();
-      giveSearchFocus();
-    }
-  };
-
-  this.enableGlobalShortcut = function() {
-    doc.body.onkeydown = onKeyDown;
-  };
-}
-
 module.exports = function() {
   var doc = window.document,
-      inputElement = doc.getElementById(SEARCH_INPUT_ID),
+      inputElement = doc.getElementById(SearchUi.DEFAULT_SEARCH_INPUT_ID),
       searchUi = new SearchUi(doc, inputElement),
-      resultsElement = doc.getElementById(SEARCH_RESULTS_ID);
+      resultsElement = doc.getElementById(SearchUi.DEFAULT_SEARCH_RESULTS_ID),
+      searchEngine = new SearchEngine();
 
   searchUi.enableGlobalShortcut();
 
@@ -111,13 +40,13 @@ module.exports = function() {
     return;
   }
 
-  return fetchIndex(window.SEARCH_BASEURL)
+  return searchEngine.fetchIndex(window.SEARCH_BASEURL)
     .then(function(searchIndex) {
-      var query = parseSearchQuery(window.location.href),
-          results = getResults(query, searchIndex);
+      var query = searchEngine.parseSearchQuery(window.location.href),
+          results = searchEngine.getResults(query, searchIndex);
       writeResults(query, doc, inputElement, resultsElement, results);
     })
     .catch(function(error) {
       console.error(error);
     });
-}();
+};
