+++ package.json
@@ -31,6 +31,6 @@
   "dependencies": {},
   "scripts": {
     "make-bundle": "browserify -g uglifyify assets/js/search.js -o assets/js/search-bundle.js",
-    "lint": "eslint assets/js/search.js lib/jekyll_pages_api_search/search.js"
+    "lint": "eslint assets/js/search.js assets/js/search-engine.js assets/js/search-ui.js lib/jekyll_pages_api_search/search.js"
   }
 }
